- add "make copy-globus_user_proxy" - add "make grid-proxy-init-force" - add "make myproxy-logon-force" - add docker-compose.override.yml.http - add docker-compose.override.yml.https.selfsigned - rename docker-compose.override.dev.yml to docker-compose.override.yml.dev - support OVERWRITEHOST OVERWRITEPROTOCOL OVERWRITEWEBROOT OVERWRITECONDADDR

+++ nextcloud/common-lib.sh
@@ -8,6 +8,7 @@ copy0()
     src="$1"
     dst="$2"
 
+    [ "$src" == "$dst" ] && exit 1
     rm -rf "$dst"
     cp -pr "$src" "$dst"
     chmod -R go-rwx "$dst"
@@ -33,4 +34,10 @@ retry_command()
         echo "Retry [$(( COUNT++ ))/${MAX_RETRY}]: $@"
     done
     return 0
+}
+
+globus_cred_ok()
+{
+    timeleft=$(${SUDO_USER} grid-proxy-info -timeleft)
+    [ "${timeleft}" -gt 0 ]
 }
